refactor(imports): add barrels and aliases
Adds barrel exports for component domains and normalizes the\nTypeScript alias configuration for '@/'.\n\nReplaces deep component imports with domain-level barrels where\nthat remains safe under Next.js App Router semantics.\n\nKeeps client-only exports out of server-imported barrels to avoid\nRSC boundary regressions and preserve existing behavior.\n\nSimplifies import paths across the codebase and prepares the\ncomponent structure for future style co-location.\n\nNo behavioral changes.
