+++ mdx-components.tsx
@@ -1,5 +1,5 @@
 import type {MDXComponents} from 'mdx/types'
-import {mdxComponents} from '@/components/blog/MDXComponents'
+import {mdxComponents} from '@/components/blog'
 
 export function useMDXComponents(components: MDXComponents): MDXComponents {
   return {
