+++ src/components/blog/BlogPostList.tsx
@@ -1,5 +1,5 @@
 import {getSortedPosts} from '@/lib/content/blog'
-import {BlogCard} from '@/components/blog/BlogCard'
+import {BlogCard} from './BlogCard'
 
 export async function BlogPostList({limit}: {limit?: number}) {
   const posts = await getSortedPosts()
