+++ src/components/blog/client.ts
@@ -1,0 +1,1 @@
+export {default as CodeHighlight} from './CodeHighlight'
