+++ src/components/contact/SuccessPopup.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {motion} from 'framer-motion'
 import {X} from 'lucide-react'
-import BlurOverlay from '@/components/contact/BlurOverlay'
+import BlurOverlay from './BlurOverlay'
 
 interface SuccessPopupProps {
   onClose: () => void
