+++ src/components/contact/index.ts
@@ -1,0 +1,1 @@
+export {default as ContactForm} from './NewContactForm'
