--- src/components/contact/index.tsx
@@ -1,1 +1,0 @@
-export {default as ContactForm} from '@/components/contact/NewContactForm'
