+++ src/components/contact/success/index.ts
@@ -1,0 +1,1 @@
+export {default as FormSuccess} from './FormSuccess'
