--- src/components/contact/success/index.tsx
@@ -1,1 +1,0 @@
-export {default as FormSuccess} from '@/components/contact/success/FormSuccess'
