+++ src/components/layout/index.ts
@@ -1,0 +1,3 @@
+export {BaseWrapper} from './BaseWrapper'
+export {default as Footer} from './Footer'
+export * from './nav'
