+++ src/components/layout/nav/index.ts
@@ -1,0 +1,1 @@
+export {Nav} from './Nav'
