--- src/components/layout/nav/index.tsx
@@ -1,1 +1,0 @@
-export {Nav} from '@/components/layout/nav/Nav'
