+++ src/components/projects/index.ts
@@ -1,0 +1,1 @@
+export {default as ProjectCard} from './ProjectCard'
