--- src/components/projects/index.tsx
@@ -1,1 +1,0 @@
-export {default as ProjectCard} from '@/components/projects/ProjectCard'
