+++ src/components/resume/index.ts
@@ -1,0 +1,1 @@
+export {default as SkillList} from './Skills'
