+++ src/components/sections/index.ts
@@ -1,0 +1,4 @@
+export * from './main'
+export {default as AboutSection} from './AboutSection'
+export {default as BlogSection} from './BlogSection'
+export {default as ProjectSection} from './ProjectSection'
