+++ src/components/sections/main/index.ts
@@ -1,0 +1,1 @@
+export {default as MainSection} from './MainSection'
