+++ src/components/shared/languages/index.ts
@@ -1,0 +1,1 @@
+export {default as LanguagesSprites} from './LanguagesSprites'
