+++ src/components/ui/index.ts
@@ -1,0 +1,2 @@
+export {default as CTA} from './CTA'
+export {default as SectionHeader} from './SectionHeader'
